Prevent integer overflows when computing configured cache_dir size.

+++ src/fs/rock/RockSwapDir.cc
@@ -276,7 +276,7 @@ Rock::SwapDir::parseSize()
     const int i = GetInteger();
     if (i < 0)
         fatal("negative Rock cache_dir size value");
-    max_size = i << 20; // MBytes to Bytes
+    max_size = static_cast<uint64_t>(i) << 20; // MBytes to Bytes
 }
 
 /// check the results of the configuration; only level-0 debugging works here
